add AMD SVM and other new opcodes to decode tables * added 24 new instructions: AMD SVM + rdtscp and Intel VMX, Westmere, Atom,   and Sandy Bridge additions * i#65: enable api/dis test to ensure I don't break the decoder   it causes ctest to run out of memory when trying to compare its large   output b/c ctest's regex matcher isn't very good: so I'm using a separate   script here to compare the output. * support setting options for standalone apps (here, api.dis test): drrun   -noinject creates .1config file, and dr_standalone_init removes it after   reading it (so no option synch support)
SVN-Revision: 476

+++ suite/tests/api/dis.c
@@ -36,6 +36,7 @@
  * Disassembles a binary file containing nothing but code.
  */
 
+#include "configure.h"
 #include "dr_api.h"
 #include <assert.h>
 
@@ -50,7 +51,7 @@ read_data(file_t f, void *drcontext)
 {
     byte sbuf[BUF_SIZE];
     byte *pc, *prev_pc;
-    int len, prev_buf_len = 0;
+    ssize_t len, prev_buf_len = 0;
 
     /* FIXME: re-run 64-bit asking for 32-bit mode */
